requirements.txt
pyfingerprint>=1.5
opencv-python>=4.5.0
numpy>=1.19.0
tensorflow>=2.6.0; platform_system != "Windows" or platform_machine != "ARM"
tflite-runtime>=2.5.0; platform_system == "Linux" and platform_machine == "ARM"
mtcnn>=0.1.0
pillow>=8.0.0
scikit-learn>=0.24.0
scipy>=1.6.0
RPi.GPIO>=0.7.0; platform_system == "Linux"
smbus>=1.1.post2; platform_system == "Linux"
RPLCD>=1.3.0; platform_system == "Linux" 